Added new functionality and alot of tests

+++ examples/smart_building/generator/energy_sensor/src/arguments.rs
@@ -5,7 +5,7 @@ use log::LevelFilter;
 #[command(version, about)]
 pub struct Arguments {
     #[arg(short = 'i', long)]
-    pub interval: f64,
+    pub interval: u64,
 
     #[arg(short = 'l', long, default_value_t = LevelFilter::Info)]
     pub level: LevelFilter
